Add react-native to project
Move project into android directory and add react files above it. I haven't
found a more sane way of structuring this, so unfortunately this is how it
has to be structured for now. I would personally like to move all of this react
stuff into the jni directory where it belongs, but for now this will do. Also
adds a .gitignore at the tld for node_modules, watchman, and react logfiles.